+++ lib/poodr_rspec/gear.rb
@@ -9,7 +9,7 @@ module PoodrRspec
     end
 
     def gear_inches
-      ratio * wheel.diameter
+      ratio * wheel.width
     end
 
     def ratio