+++ lib/poodr_rspec/wheel.rb
@@ -6,7 +6,7 @@ module PoodrRspec
       @tire = tire
     end
 
-    def diameter
+    def width
       rim + (tire * 2)
     end
   end