+++ spec/support/shared_examples/a_diameterizable_interface.rb
@@ -1,0 +1,3 @@
+RSpec.shared_examples 'a diameterizable interface' do
+  it { expect(diameterizable).to respond_to(:width) }
+end