+++ test/poodr_rspec/wheel_test.rb
@@ -1,11 +1,16 @@
 require 'test_helper'
+require 'support/modules/diameterizable_interface_test'
 
 module PoodrRspec
   class WheelTest < MiniTest::Test
-    def test_calculate_diameter
-      wheel = Wheel.new(26, 1.5)
+    include DiameterizableInterfaceTest
 
-      assert_in_delta(29, wheel.diameter, 0.01)
+    def setup
+      @wheel = @object = Wheel.new(26, 1.5)
+    end
+
+    def test_calculate_diameter
+      assert_in_delta(29, @wheel.width, 0.01)
     end
   end
 end