+++ test/support/modules/diameterizable_interface_test.rb
@@ -1,0 +1,5 @@
+module DiameterizableInterfaceTest
+  def test_implemets_the_diameterizable_interface
+    assert_respond_to(@object, :width)
+  end
+end